# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Moxxy Agentic Framework API
  version: 0.1.0
  description: Local-first agentic framework REST API
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "mox_*"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    TemplateUpdate:
      type: object
      properties:
        content:
          type: string
          description: Full TEMPLATE.md content with YAML frontmatter
      required: [content]
    TemplateResponse:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        version:
          type: string
        tags:
          type: array
          items:
            type: string
        body:
          type: string
          description: Instruction content (post-frontmatter)
...
paths:
...
  /v1/templates/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Templates]
      summary: Get template by slug
      responses:
        '200':
          description: Template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Templates]
      summary: Update a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateUpdate'
      responses:
        '200':
          description: Template updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '400':
          description: Invalid content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Templates]
      summary: Delete a template
      description: Deletes the template and clears it from any agents that reference it.
      responses:
        '200':
          description: Template deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                  slug:
                    type: string
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...